Add the ability to switch between public and private IP Addresses
Many providers have both public and private IP Addresses that can
be assigned to instances.

This patch allows you to switch between the two through the use of
the --ip_type flag which can be set to either "public" or "private".

This is especially useful when running in a Virtual Private Cloud
(VPC) inside AWS as all instances inside the VPC usually have a
private address not a public one.

The flag is optional and defaults to "public" so therefore should
not break existing behaviour.